\name{jetairplane}
\alias{jetairplane}
\docType{data}
\title{
Number of failures of Boeing Jets
}
\description{
A data set containing number of failures for air conditioning systems of jet airplane data.
}
\usage{data("jetairplane")}
\format{
  A data frame with 212 observations on the following variable.
  \describe{
    \item{\code{failures}}{a numeric vector}
  }
}

\source{
Exponentiated Kumaraswamy-Dagum distribution with applications to income and lifetime data
}

\examples{
data(jetairplane)
## maybe str(jetairplane) ; plot(jetairplane) ...
}
\keyword{datasets}
